fix stats-service

+++ stats/stats-server/src/main/java/ru/practicum/ewm/stats/domain/ViewStats.java
@@ -13,5 +13,5 @@ import lombok.experimental.FieldDefaults;
 public class ViewStats {
     String app;
     String uri;
-    int hits;
+    long hits;
 }
