+++ stats/stats-server/src/main/java/ru/practicum/ewm/stats/service/mapping/ViewStatsMapping.java
@@ -5,6 +5,6 @@ import ru.practicum.ewm.stats.domain.ViewStats;
 
 public class ViewStatsMapping {
     public static ViewStatsDto toViewStatsDto(ViewStats hits) {
-        return new ViewStatsDto(hits.getApp(), hits.getUri(), hits.getHits());
+        return new ViewStatsDto(hits.getApp(), hits.getUri(), (int) hits.getHits());
     }
 }
